refactor: split optimizeSvg logic and types
