+++ src/types/svg.d.ts
@@ -1,0 +1,1 @@
+type SvgOptions = Partial<UnionToIntersection<import('svgo').PluginConfig>>
