+++ tsconfig.json
@@ -9,6 +9,7 @@
     "noUnusedLocals": true,
     "outDir": "dist",
     "strict": true,
+    "typeRoots": ["./node_modules/@types", "./src/types"],
     "sourceMap": true,
     "target": "es2018"
   }
